(DataView) Moves save view button to primary buttons row to align with Apex design system.
