+++ containers/mosaic/src/components/DataView/DataViewViewControls/DataViewViewControls.tsx
@@ -1,44 +1,16 @@
-import React, { useMemo } from "react";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+import React from "react";
 
 import ButtonRow from "../../ButtonRow";
-import Button from "../../Button";
 
-import { useMosaicTranslation } from "@root/i18n";
 import type { DataViewViewControlsProps } from "./DataViewViewControlsTypes";
-import type { MenuItemProps } from "@root/components/MenuItem";
 import testIds from "@root/utils/testIds";
 import { DataViewViewButton } from "./DataViewViewButton";
 import { DataViewViewListWrapper } from "./DataViewViewListWrapper";
 
 function DataViewViewControls({
 	currentView,
-	onViewSave,
-	onViewSaveAs,
 	onViewList,
 }: DataViewViewControlsProps) {
-	const { t } = useMosaicTranslation();
-
-	const viewMenuItems = useMemo<MenuItemProps[]>(() => {
-		const listItems: MenuItemProps[] = [];
-
-		if (onViewSave) {
-			listItems.push({
-				label: t("mosaic:DataView.overwrite_current_view"),
-				onClick: onViewSave,
-			});
-		}
-
-		if (onViewSaveAs) {
-			listItems.push({
-				label: t("mosaic:DataView.save_as_new_view"),
-				onClick: onViewSaveAs,
-			});
-		}
-
-		return listItems;
-	}, [onViewSave, onViewSaveAs, t]);
-
 	return (
 		<div data-testid={testIds.DATA_VIEW_VIEW_CONTROL}>
 			<div className="right">
@@ -53,14 +25,6 @@ function DataViewViewControls({
 							currentView={currentView}
 						/>
 					)}
-					<Button
-						mIcon={ExpandMoreIcon}
-						iconPosition="right"
-						label={t("mosaic:DataView.save_view")}
-						variant="text"
-						intent="info"
-						menuItems={viewMenuItems}
-					/>
 				</ButtonRow>
 			</div>
 		</div>
