+++ containers/mosaic/src/components/DataView/DataViewViewControls/DataViewViewControlsTypes.ts
@@ -8,8 +8,6 @@ import type { DataViewProps } from "../DataViewTypes";
 
 export type DataViewViewControlsProps = Pick<DataViewProps,
 	| "currentView"
-	| "onViewSave"
-	| "onViewSaveAs"
 	| "onViewList"
 >;
